# core: use less memory during reorgs