One thing that happened on the goerli-shadow-fork-1 testnet was that a really long reorg happened.
```
WARN [03-30|09:20:00.399] Long reorg                               newchain length=6,598,237 from=6,598,237 to=1
```
This means that the whole chain gets reorged in. This can only happen if we have the whole chain in the database but not marked canonical. I suspect that we started a sync which almost succeeded (the node had a lot of unclean shutdowns etc) so that we had all the blocks in the database and then we got a fcu which marks a post merge block as canonical. It also starts the reorg procedure. The problem now was that the reorg procedure took (understandably) too much memory and the node ran OOM. Would be good if @karalabe could double check my reasoning here or if you suspect anything different going on why the chain was not marked as canonical.


This PR significantly reduces the memory consumption of the reorg procedure. ~~It does increase the amount of database access by 2x though as every block has to be retrieved from the db twice.~~ 
It might make sense to have this though as post-merge longer reorgs can happen more frequently. 
Blocks are imported as non-canonical via `engine_newPayload` and are marked later on as canonical with `engine_forkChoiceUpdated`. In times of non-finalization it could be that a lot of blocks are imported as non-canonical and the reorg procedure is only called after a long time. So it might make sense to reduce the memory consumption of the reorg procedure