@MariusVanDerWijden 2022-05-23
Just changed the PR quite a bit in order to reduce the disk lookups etc. 
Now this PR is faster than master, with only slightly more overall allocations.

Master: BenchmarkReorg-24    	   31770	     35593 ns/op	   17308 B/op	     278 allocs/op
This PR: BenchmarkReorg-24    	    31952	     34184 ns/op	   17612 B/op	     278 allocs/op

I've created a test that create two 10.000 block long chains with 500 transactions per block and then reorged the chain and dumped the memory consumption multiple times during the reorg method to see the peak memory consumption.

This PR:
Alloc = 9696 MiB        TotalAlloc = 173439 MiB Sys = 15208 MiB NumGC = 104
Alloc = 9594 MiB        TotalAlloc = 176626 MiB Sys = 16267 MiB NumGC = 105
Alloc = 10893 MiB       TotalAlloc = 186432 MiB Sys = 18410 MiB NumGC = 106
Time needed: 48.775932281s⏎ 

Master:
Alloc = 9026 MiB        TotalAlloc = 172549 MiB Sys = 14871 MiB NumGC = 104
Alloc = 14228 MiB       TotalAlloc = 182057 MiB Sys = 15711 MiB NumGC = 105
Alloc = 10969 MiB       TotalAlloc = 185151 MiB Sys = 17468 MiB NumGC = 106
Time needed: 48.253782763s⏎ 

As you can see, the peak memory consumption goes down from 14 GB to 10GB.